feat: move rpc / client base / server base to a more proper path
