+++ robots/libero/tools.py
@@ -12,8 +12,8 @@ from robots.libero.env_client import LiberoEnvClient
 from rpent.tools.state import EnvState, StepRecord
 from rpent.tools.toolkit import readonly
 from rpent.utils.logging import get_logger
-from rpent.utils.sam3_client import Sam3Client
-from rpent.utils.vla_client import VLAClient
+from rpent.robots.components.sam3_client import Sam3Client
+from rpent.robots.components.pi05_vla_client import VLAClient
 
 logger = get_logger("libero")
