+++ robots/robotwin/vla_server.py
@@ -13,7 +13,7 @@ if __package__ in (None, ""):
     sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
 from robots.robotwin.robot_spec import MODEL_SPEC, vla_runtime_contract
-from rpent.tools.vla_facade_base import BaseVLAFacade
+from rpent.robots.components.vla_facade_base import BaseVLAFacade
 from rpent.utils.daemon import watch_parent_death
